Move common cmds into e2e dir
