+++ cypress/support/commands/e2e/index.js
@@ -25,7 +25,8 @@
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
 
 import '@testing-library/cypress/add-commands';
-import './schema'
+import './common';
+import './schema';
 import './api';
 import './auth';
 import './repository';
